Rename dev_runtime to browser_runtime
